% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cofmat}
\alias{Cofmat}
\title{Cofmath values calculation
Transforming the vector c=(c_1,...,c_Ld) to a d*L matrix of coeficients.}
\usage{
Cofmat(d, L, cx)
}
\description{
Cofmath values calculation
Transforming the vector c=(c_1,...,c_Ld) to a d*L matrix of coeficients.
}
