% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftscor.R
\name{cor.fts}
\alias{cor.fts}
\title{Correlation for Functional Time Series Objects}
\usage{
cor.fts(Y1, Y2)
}
\arguments{
\item{Y1}{a functional time series object of class \code{fts}.}

\item{Y2}{a functional time series object of class \code{fts}.}
}
\value{
the correlation between functional time series.
}
\description{
This function finds the correlation between univarite or multivariate functional time series objects
}
\seealso{
\code{\link{fts}}
}
