% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtscor.R
\name{cor.funts}
\alias{cor.funts}
\title{Correlation for Functional Time Series (funts) Objects}
\usage{
cor.funts(Y1, Y2)
}
\arguments{
\item{Y1}{an object of class \code{\link{funts}}.}

\item{Y2}{an object of class \code{\link{funts}}.}
}
\value{
A scalar representing the correlation between \code{\link{funts}} objects.
}
\description{
Find the correlation between univariate or multivariate functional time series
(\code{\link{funts}}) objects.
}
\examples{
\dontrun{
require(Rfssa)
data("Montana")
## Raw image data
NDVI <- Montana[, 1]
EVI <- Montana[, 2]
cor.funts(NDVI, EVI)
}

}
\seealso{
\code{\link{funts}}
}
