% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deathrate.R
\docType{data}
\name{Deathrate}
\alias{Deathrate}
\title{The Age-specific death rates dataset.}
\format{A data frame with 10434 rows and 5 variables:
\describe{
  \item{Year}{The year}
  \item{Age}{The age}
  \item{Female}{The female death rate}
  \item{Male}{The male death rate}
  \item{Total}{The total death rate}
}}
\source{
\url{http://www.mortality.org/}
}
\usage{
Deathrate
}
\description{
The death rates for years 1921 to 2014 and age range 0 to 100.
}
\keyword{datasets}
