% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts_methods.R
\name{-.funts}
\alias{-.funts}
\title{Subtract `funts` Objects or a `funts` Object and a Scalar}
\usage{
\method{-}{funts}(obj1, obj2 = NULL)
}
\arguments{
\item{obj1}{an object of class `funts`, representing the first `funts` object.}

\item{obj2}{an object of class `funts` or a scalar.}
}
\description{
This function allows subtraction between two `funts` (functional time series) objects or
between a `funts` object and a scalar. It returns the resulting difference.
}
\examples{
\dontrun{
data("Callcenter")
y <- Callcenter
print(1 - y)
}
}
\seealso{
\code{\link{funts}}
}
