% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fforecast.r
\name{fforecast}
\alias{fforecast}
\title{Functional Singular Spectrum Analysis R-Forecasting and V-Forecasting}
\usage{
fforecast(U, groups = list(c(1)), h = 1, method = "recurrent", tol = 10^-3)
}
\arguments{
\item{U}{an object of class \code{\link{fssa}} that holds the decomposition}

\item{groups}{a list of length one containing a vector of the indices of the left singular functions to be used in the forecast}

\item{h}{an integer that specifies the forecast horizon}

\item{method}{a character string specifying the type of forecasting to perform}

\item{tol}{a double specifying the amount of tolerated error in the approximation of the Neumann series leveraged in both forecasting algorithms
\dontrun{
# Univariate FSSA Forecasting
require(fda)
require(Rfssa)
data("Callcenter")
## Define functional objects
D <- matrix(sqrt(Callcenter$calls),nrow = 240)
N <- ncol(D)
time <- seq(ISOdate(1999,1,1), ISOdate(1999,12,31), by="day")
K <- nrow(D)
u <- seq(0,K,length.out=K)
d <- 23
basis <- create.bspline.basis(c(min(u),max(u)),d)
Ysmooth <- smooth.basis(u,D,basis)
## Define functional time series
Y <- fts(Ysmooth$fd,time = time)
plot(Y,xlab="Time (6 minutes aggregated)", ylab="Square root of call numbers",main = "Call Center Data")
## Perfrom FSSA decomposition
L <- 28
U <- fssa(Y,L)
groups <- list(1:7)
## Perfrom FSSA R-forecast and FSSA V-forecast
pr_R <- fforecast(U = U, groups = groups, h = 50, method = "recurrent", tol = 10^-3)
plot(pr_R,xlab="Time (6 minutes aggregated)", ylab="Square root of call numbers",main = "Recurrent Forecast")
pr_V <- fforecast(U = U, groups = groups, h = 50, method = "vector", tol = 10^-3)
plot(pr_V,xlab="Time (6 minutes aggregated)", ylab="Square root of call numbers",main = "Vector Forecast")


# Multivariate FSSA Forecasting
NDVI=Jambi$NDVI
EVI=Jambi$EVI
time <- Jambi$Date
## Kernel density estimation of pixel intensity
D0_NDVI <- matrix(NA,nrow = 512, ncol = 448)
D0_EVI <- matrix(NA,nrow =512, ncol = 448)
for(i in 1:448){
 D0_NDVI[,i] <- density(NDVI[,,i],from=0,to=1)$y
 D0_EVI[,i] <- density(EVI[,,i],from=0,to=1)$y
}
## Define functional objects
d <- 11
basis <- create.bspline.basis(c(0,1),d)
u <- seq(0,1,length.out = 512)
y_NDVI <- smooth.basis(u,as.matrix(D0_NDVI),basis)$fd
y_EVI <- smooth.basis(u,as.matrix(D0_EVI),basis)$fd
y=list(y_NDVI,y_EVI)
## Define functional time series
Y <- fts(y,time=time)
plot(Y,xlab="NDVI Value", ylab=c("Density","Density"),main = "NDVI Densities")
## Perform MFSSA decompsosition
U <- fssa(Y,L=45)
groups <-list(1:4)
## Perform MFSSA forecasting
pr_R_M <- fforecast(U = U, groups = groups, h = 56, method = "recurrent", tol = 10^-3)
plot(pr_R_M,xlab="NDVI Value", ylab=c("Density","Density"),main = "Recurrent Forecast")
pr_V_M <- fforecast(U = U, groups = groups, h = 56, method = "vector", tol = 10^-3)
plot(pr_V_M,xlab="NDVI Value", ylab=c("Density","Density"),main = "Vector Forecast")

}}
}
\value{
An object of class \code{\link{fts}}
}
\description{
This function performs FSSA and MFSSA R-forecasting or V-forecasting of functional time series (\code{\link{fts}})
}
