% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpredinterval.R
\name{fpredinterval}
\alias{fpredinterval}
\title{FSSA Forecasting Bootstrap Prediction Interval}
\usage{
fpredinterval(
  Y,
  O,
  L,
  ntriples,
  Bt,
  h = 1,
  alpha = 0.05,
  method = "recurrent",
  tol = 10^-3
)
}
\arguments{
\item{Y}{An object of class \code{\link{funts}}.}

\item{O}{A positive integer specifying the training set size.}

\item{L}{A positive integer specifying the window length.}

\item{ntriples}{The number of eigentriples to use to perform the forecasts.}

\item{Bt}{A positive integer specifying the number of bootstrap samples.}

\item{h}{An integer that specifies the forecast horizon.}

\item{alpha}{A double between zero and one specifying the significance level.}

\item{method}{A character string specifying the type of forecasting to perform either \code{"recurrent"} or \code{"vector"}.}

\item{tol}{A double specifying the amount of tolerated error in the approximation of the matrix that corresponds with the operator formed using a Neumann series leveraged in both forecasting algorithms.}
}
\value{
A list of numeric vectors where the first vector contains the discretely sampled point forecast, the second contains the discretely sampled lower bound of the prediction interval, and the third contains the discretely sampled upper bound of the prediction interval.
}
\description{
This function calculates the bootstrap prediction interval for functional singular spectrum analysis (FSSA) forecasting predictions of univariate functional time series (\code{\link{funts}}) observed over a one-dimensional domain.
}
\examples{
\dontrun{
require(ggplot2)
data("Callcenter")
# Calculate prediction intervals

pred_interval <- fpredinterval(Y = Callcenter, O = 310, L = 28, ntriples = 7, Bt = 10000, h=3, alpha = 0.05, method = "recurrent")

# Plot the forecast and prediction interval using ggplot

fssa_forecast <- pred_interval$forecast
fssa_forecast_lower <- pred_interval$lower
fssa_forecast_upper <- pred_interval$upper
ggplot(data = data.frame(cbind(fssa_forecast,fssa_forecast_lower,fssa_forecast_upper)),aes(1:K,fssa_forecast))+theme_bw()+
geom_ribbon(aes(ymin=fssa_forecast_lower,ymax=fssa_forecast_upper),fill="grey90")+
  geom_line(aes(1:K,fssa_forecast_upper),color="black",size=0.5)+
  geom_line(aes(1:K,fssa_forecast_lower),color="black",size=0.5)+
  geom_line(aes(1:K,fssa_forecast),color="red2",size=1)+
  scale_x_continuous(name = "Time (6 minutes aggregated)", breaks=c(1,60,120,180,240),
                     labels = c("00:00","06:00","12:00","18:00","24:00"),)+
  scale_y_continuous(name="Sqrt of Call Numbers")+
  theme(axis.text.x=element_text(size=15),
        axis.title.x=element_text(size=18))+
  theme(axis.text.y=element_text(size=15),
        axis.title.y=element_text(size=18))+ggtitle("Prediction Intervals for Jan. 3, 2000")+
  theme(plot.title = element_text(hjust = 0.5,size=20))
}

}
