% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freconstruct.r
\name{freconstruct}
\alias{freconstruct}
\title{Reconstruction sage of FSSA}
\usage{
freconstruct(U, group = as.list(1L:10L))
}
\arguments{
\item{U}{an object of class \code{\link{fssa}}}

\item{group}{a list of numeric vectors, each vector includes indices of such elementary components
of a group used for reconstruction.}
}
\value{
a named list of reconstructed functional time series in each groups and
a numeric vector of eigenvalues.
}
\description{
This is a function for the reconstructing functional time series from
functional singulare spectrum objects (including Grouping and
Hankelization steps). The output is a list of functional time series corresponds to each group.
}
\seealso{
\code{\link{fssa}}
}
