% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfssa.R, R/ftsplot.r
\name{ftsplot}
\alias{ftsplot}
\title{3D fts Plots}
\usage{
ftsplot(x, y, X, type = 2, zlab = NULL, xlab = NULL, ylab = NULL,
  space = 0.1, main = NULL)

ftsplot(x, y, X, type = 2, zlab = NULL, xlab = NULL, ylab = NULL,
  space = 0.1, main = NULL)
}
\description{
3D fts Plots

3D fts Plots
}
