% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.funfts}
\alias{is.funfts}
\title{Check if an object is of class 'funfts'}
\usage{
is.funfts(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
TRUE if the object is of class 'funfts', FALSE otherwise.
}
\description{
Check if an object is of class 'funfts'
}
