% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.funts}
\alias{is.funts}
\title{Check if an object is of class 'funts'}
\usage{
is.funts(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
TRUE if the object is of class 'funts', FALSE otherwise.
}
\description{
Check if an object is of class 'funts'
}
\examples{
\dontrun{
data("Callcenter")
is.funts(Callcenter)
}
}
