% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jambi.R
\docType{data}
\name{Jambi}
\alias{Jambi}
\title{Jambi MODIS Data}
\format{A list which contains two 33 by 33 by 448 arrays where one array is for NDVI image data and the other is for EVI image data. The list also contains a date vector of length 448
\describe{
  \item{Days 1 - 448}{Pixel intensity with values between zero and one}
} @references
\enumerate{
\item
 Lambin, E., Geist, H., Lepers, E. (1999).
 Dynamics of Land-Use and Land-Cover Change in Tropical Regions
 \emph{Annual Review of Environment and Resources}, 205-244.


}
The raw data can be found at @source \url{https://lpdaac.usgs.gov/products/mod13q1v006/}}
\usage{
Jambi
}
\description{
This data set contains the normalized difference vegetation index (NDVI) and enhanced vegetation index (EVI) image data from NASA’s MODerate-resolution Imaging Spectroradiometer (MODIS) with global coverage at 250 m^2.
The goal of the study is to collect raw image data of Jambi Province, Indonesia.
Indonesia manages various forested land utilizations, for instance natural forest and plantations.
In the past, natural forest had been exploited throughout the country.
Greater criticisms on forest exploitation lead to a moratorium which needs to be
monitored frequently. Assessment of woody vegetation could be taken using field surveys or
remote sensing. Season is probably the most intriguing factor in vegetative land cover,
especially in long-term land cover changes (Lambin, 1999).
The data was gathered starting in 2000-02-18 and ending in 2019-07-28 every 16 days.
The data set contains the image data for NDVI, EVI, and a date object which gives the date of when each image was taken.
}
\keyword{datasets}
