% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts_methods.R
\name{length.funts}
\alias{length.funts}
\title{Length of Functional Time Series}
\usage{
\method{length}{funts}(obj)
}
\arguments{
\item{obj}{an object of class "funts" .}
}
\description{
Returns the length of a "funts" object.
}
