% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_github_data.R
\name{loadCallcenterData}
\alias{loadCallcenterData}
\title{Load Callcenter Data from GitHub Repository}
\format{
A dataframe with 87,600 rows and 5 variables:
\describe{
  \item{calls}{The number of calls in a 6-minute aggregated interval.}
  \item{u}{A numeric vector indicating the aggregated interval.}
  \item{Date}{Date and time when the call counts were recorded.}
  \item{Day}{Weekday associated with Date.}
  \item{Month}{Month associated with Date.}
}
}
\usage{
loadCallcenterData()
}
\description{
This function loads the Callcenter dataset from the Rfssa_dataset repository hosted on GitHub
(https://github.com/haghbinh/dataset/Rfssa_dataset). Hosting datasets on GitHub instead of
including them in the Rfssa R package saves a significant amount of space.
The Callcenter dataset represents a small call center for an anonymous bank, providing the exact time
of calls connected to the center from January 1 to December 31 in the year 1999. The data is aggregated
into time intervals, with each element of the data matrix representing call volume during a specific
6-minute interval on a particular day.
}
\examples{
\dontrun{
loadCallcenterData()
}

}
\references{
\enumerate{
\item
Brown, L., Gans, N., Mandelbaum, A., Sakov, A., Shen, H., Zeltyn, S., & Zhao, L. (2005).
Statistical analysis of a telephone call center: A queueing-science perspective.
\emph{Journal of the American Statistical Association}, \strong{100}(469), 36-50.
\item
Shen, H., & Huang, J. Z. (2005).
Analysis of call center arrival data using singular value decomposition.
\emph{Applied Stochastic Models in Business and Industry}, \strong{21}(3), 251-263.
\item
Huang, J. Z., Shen, H., & Buja, A. (2008).
Functional principal components analysis via penalized rank one approximation.
\emph{Electronic Journal of Statistics}, \strong{2}, 678-695.
\item
Maadooliat, M., Huang, J. Z., & Hu, J. (2015).
Integrating data transformation in principal components analysis.
\emph{Journal of Computational and Graphical Statistics}, \strong{24}(1), 84-103.
}
}
\seealso{
\code{\link{Callcenter}}, \code{\link{funts}}
}
