% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_github_data.R
\name{loadMontanaData}
\alias{loadMontanaData}
\title{Load Montana Data from GitHub Repository}
\usage{
loadMontanaData()
}
\description{
This function is used to load the Montana dataset from the Rfssa package hosted on GitHub at
https://github.com/haghbinh/Rfssa. Hosting datasets on GitHub instead of including them
in the Rfssa R package saves a significant amount of space.
}
\examples{
\dontrun{
loadMontanaData()
}

}
