% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_github_data.R
\name{loadTempData}
\alias{loadTempData}
\title{Load Austin and Utqiagvik Temperature Data from GitHub Repository}
\format{
Containing two matrix objects:
\describe{
  \item{Austin Temperature Data}{A 24 by 607 matrix of discrete samplings of intraday hourly temperature curves, one day per month.}
  \item{Utqiagvik Temperature Data}{A 24 by 607 matrix of discrete samplings of intraday hourly temperature curves, one day per month.}
}
}
\usage{
loadTempData()
}
\description{
This function loads the Austin/Utqiagvik Temperature dataset from a GitHub repository hosted at
https://github.com/haghbinh/dataset/Rfssa_dataset. Hosting datasets on GitHub
rather than including them in the Rfssa R package conserves storage space.
}
\details{
The Austin/Utqiagvik Temperature dataset contains intraday hourly temperature curves
measured in degrees Celsius from January 1973 to July 2023, recorded once per month.
}
\examples{
\dontrun{
loadTempData()
str(austin)
str(utqiagvik)
}


}
