% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTSclass.R
\name{mfts}
\alias{mfts}
\title{Multivariate Functional Time Series (mfts) Class}
\usage{
mfts(X, basisobj, argval = NULL, method = "data", start = 1, end = NULL)
}
\arguments{
\item{X}{A matrix, three-dimensional array, or a list containing matrix or array objects. When `method="data"`, it represents
the set of curve values at discrete sampling points or argument values. When `method="coefs"`, `X` specifies the coefficients
corresponding to the basis system defined in 'basisobj'. If `X` is a list, it defines a multivariate FTS, with each element
being a matrix or three-dimensional array object. In matrix objects, rows correspond to argument values, and columns correspond
to the length of the FTS. In three-dimensional array objects, the first and second dimensions correspond to argument values,
and the third dimension to the length of the FTS.}

\item{basisobj}{An object of class `basisfd`, a matrix of empirical basis, or a list of `basisfd` or empirical basis objects.
For empirical basis, rows correspond to basis functions, and columns correspond to grid points.}

\item{argval}{A vector list of length `p`. Each entry in the list should either be a numeric value or a list of numeric
elements, depending on the dimension of the domain the variable is observed over.}

\item{method}{Determines the type of the `X` matrix: "coefs" or "data."}

\item{start}{The time of the first observation. It can be a single number or an object of classes `Date`, `POSIXct`, or `POSIXt`,
specifying a natural time unit.}

\item{end}{The time of the last observation, specified in the same way as `start`.}
}
\description{
The `mmfts` class is designed to encapsulate functional time series objects, both univariate (FTS) and multivariate (MFTS).
It allows you to create `mfts` objects with specified basis systems. You can create an `mfts` object either by providing discrete
observed samples of curves and applying smoothing using a basis method or by supplying the set of basis coefficients.
`mfts` objects can handle two-dimensional functional data and different dimensions for multivariate FTS.
}
\note{
Refer to \code{\link{fssa}} for examples on how to use this function.
}
\seealso{
\code{\link{fssa}}
}
