% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts_methods.R
\name{plot.funts}
\alias{plot.funts}
\title{Plot Functional Time Series (funts) Data}
\usage{
\method{plot}{funts}(
  obj,
  npts = 100,
  obs = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  type = "l",
  lty = 1,
  ...
)
}
\arguments{
\item{obj}{an object of class \code{funts}.}

\item{npts}{number of grid points for the plots.}

\item{obs}{observation number (for two-dimensional domains).}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main title for the plot.}

\item{type}{type of plot ("l" for line, "p" for points, etc.).}

\item{lty}{line type (1 for solid, 2 for dashed, etc.).}

\item{...}{additional graphical parameters passed to plotting functions.}
}
\description{
Create visualizations of Functional Time Series (funts) data. This function supports
plotting funts data with one-dimensional or two-dimensional domains.
}
\details{
The function allows plotting funts data with one-dimensional domains as line plots and funts data
with two-dimensional domains as image plots.
}
\examples{
# Example with one-dimensional domain
data("Callcenter")
plot(Callcenter,lwd=2, npts = 200, col = "deepskyblue4",
    main = "Call Center Data",
    xlab = "Time (6 minutes aggregated)",
    ylab = "Sqrt of Call Numbers")

# Example with two-dimensional domain
data("Montana")
plot(Montana, obs = 2,
    main = c("Temperature Curves", "NDVI Images,"),
    xlab = c("Time", "Longitude"),
    ylab = c("Normalized Temperature (\u00B0C)", "Latitude"))

}
\seealso{
\code{\link{funts}}, \code{\link{Callcenter}}, \code{\link{Montana}}
}
