% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfssa-package.R
\docType{package}
\name{Rfssa-package}
\alias{Rfssa}
\alias{Rfssa-package}
\title{Rfssa: Functional Singular Spectrum Analysis}
\description{
Methods and tools for implementing functional singular spectrum analysis and related techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haghbinh/Rfssa}
}

}
\author{
\strong{Maintainer}: Hossein Haghbin \email{haghbin@pgu.ac.ir} (\href{https://orcid.org/0000-0001-8416-2354}{ORCID})

Authors:
\itemize{
  \item Jordan Trinka \email{jordantrinka4@hotmail.com}
  \item Seyed Morteza Najibi \email{mor.najibi@gmail.com}
  \item Mehdi Maadooliat \email{mehdi.maadooliat@mu.edu} (\href{https://orcid.org/0000-0002-5408-2676}{ORCID})
}

}
\keyword{internal}
