% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFSSA.R
\name{server.fssa}
\alias{server.fssa}
\title{Shiny app server function}
\usage{
server.fssa(input, output, clientData, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{clientData}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function
}
