% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftsarithmetic.R
\name{[.fts}
\alias{[.fts}
\title{Indexing into Functional Time Series}
\usage{
\method{[}{fts}(Y, i = "index")
}
\arguments{
\item{Y}{A Univariate of Multivariate functional time series.}
}
\value{
An object of class fts.

\item{out}{A functional time series}
}
\description{
[ is an S3 method that lets you index into a functional time series
}
\examples{

station_data = read.csv(file = "C:/Users/jorda/Dropbox/Jordan_Trinka_Research/MFSSA/Data/Pedestrian/Pedestrian_2015/January_2015.csv",header = TRUE)
library(fda)
library(Rfssa)
x=1:24
y=station_data[1:nrow(station_data),33]
nr=24
d=11
basis <- create.fourier.basis(c(0,1),d)
u=seq(0,1,length.out = 24)
Y=list()
for(j in 1:ncol(y)){

 Y[[j]]=smooth.basis(argvals = u,y = sqrt(matrix(data=y[,j],nr=nr)),fdParobj = basis)$fd


}

Y=fts(Y)


plot(Y[3:5])





}
