% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfts_methods.R
\name{[.funts}
\alias{[.funts}
\title{Indexing into Functional Time Series}
\usage{
\method{[}{funts}(obj, i = "index")
}
\arguments{
\item{obj}{An object of class \code{\link{funts}}.}

\item{i}{An index or indices specifying the subsets to extract.}
}
\description{
An indexing method for functional time series (\code{\link{funts}}) objects.
}
\details{
This function allows you to extract specific subsets of a functional time series
based on the provided indices. You can specify which subsets you want to extract
from the functional time series.
}
\seealso{
\code{\link{funts}}
}
