% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts_methods.R
\name{*.funts}
\alias{*.funts}
\title{Scalar Multiplication of Functional Time Series (funts)}
\usage{
\method{*}{funts}(obj1, obj2)
}
\arguments{
\item{obj1}{an object of class \code{\link{funts}} or a scalar value.}

\item{obj2}{an object of class \code{\link{funts}} or a scalar value.}
}
\value{
An object of class \code{\link{funts}} representing the result of scalar multiplication.
}
\description{
Perform scalar multiplication of a Functional Time Series (funts) object by either another funts object or a scalar value.
}
\details{
This function allows element-wise multiplication of a Functional Time Series (funts) object by another funts object or scalar value.
}
\seealso{
\code{\link{funts}}
}
