% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFSSA.R
\docType{data}
\name{ui.fssa}
\alias{ui.fssa}
\title{Shiny app server object}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
ui.fssa
}
\description{
Shiny app server object
}
\keyword{datasets}
