% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callcenter.R
\docType{data}
\name{Callcenter}
\alias{Callcenter}
\title{Number of calls for a bank.}
\format{A data frame with 87600 rows and 5 variables:
\describe{
  \item{calls}{The number of calls in 6 minutes aggregated interval.}
  \item{u}{a numeric vector to show the aggregated interval}
  \item{Date}{Date time when the calls counts are recorded}
  \item{Day}{Weekday associated with Date}
  \item{Month}{Month associated with Date}
}}
\usage{
Callcenter
}
\description{
This dataset is a small call center for an anonymous bank (Brown et al.,
2005). This dataset provides the exact time of the calls that were connected to
the call center between January 1 and December 31 in the year 1999.
The data are aggregated  into time intervals to obtain a data matrix. More
precisely, the i'th, row jth column of the data matrix contains the call volume
during the jth time interval on day i. This data set has been analyzed in several
prior studies; e.g. Brown et al. (2005), Huang et al. (2008), Shen and Huang
(2008) and Maadooliat et al. (2015). Here, the data are aggregated  into time
intervals 6 minutes.
}
\references{
\enumerate{
  \item
  Huang, J. Z., Shen, H., & Buja, A. (2008).
  Functional principal components analysis via
  penalized rank one approximation. \emph{Electronic Journal
  of Statistics}, 2, 678-695.
  \item
  Maadooliat, M., Huang, J. Z., & Hu, J. (2015).
  Integrating data transformation in principal
  components analysis. \emph{Journal of Computational and
  Graphical Statistics}, 24(1), 84-103.
}
}
\keyword{datasets}
