% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftscor.R
\name{cor.fts}
\alias{cor.fts}
\title{Correlation for Functional Time Series Objects}
\usage{
cor.fts(Y1, Y2)
}
\arguments{
\item{Y1}{a functional time series object of class \code{fts}.}

\item{Y2}{a functional time series object of class \code{fts}.}
}
\value{
the correlation between functional time series.
}
\description{
This function finds the correlation between univarite or multivariate functional time series objects
}
\examples{

\dontrun{
library(fda)
library(Rfssa)
d <- 11
basis <- create.bspline.basis(c(0,1),d)
u <- seq(0,1,length.out = 512)
y_1 <- smooth.basis(u,as.matrix(NDVI),basis)$fd
y_2 <- smooth.basis(u,as.matrix(EVI),basis)$fd
Y1=fts(y_1)
Y2=fts(y_2)
cor.fts(Y1,Y2)
}

}
\seealso{
\code{\link{fts}}
}
