% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftsarithmetic.R
\name{-.fts}
\alias{-.fts}
\title{Subtraction of Functional Time Series}
\usage{
\method{-}{fts}(Y1, Y2)
}
\arguments{
\item{Y1}{A Univariate or Multivariate functional time series or scalar.}

\item{Y2}{A Univariate or Multivariate functional time series or scalar.}
}
\value{
An object of class fts.

\item{out}{A functional time series}
}
\description{
- is an S3 method that lets you perform functional time series subtraction and scalar subtraction
}
\examples{

station_data = read.csv(file = "C:/Users/jorda/Dropbox/Jordan_Trinka_Research/MFSSA/Data/Pedestrian/Pedestrian_2015/January_2015.csv",header = TRUE)
library(fda)
library(Rfssa)
x=1:24
y=station_data[1:nrow(station_data),33:34]
nr=24
d=11
basis <- create.fourier.basis(c(0,1),d)
u=seq(0,1,length.out = 24)
Y=list()
for(j in 1:ncol(y)){

 Y[[j]]=smooth.basis(argvals = u,y = sqrt(matrix(data=y[,j],nr=nr)),fdParobj = basis)$fd


}

Y=fts(Y)

Y1=fts(Y[[1]])

Y2=fts(Y[[2]])

plot(Y1 - Y2)

plot(2 - Y2)

plot(Y1 - 3)


}
