% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fforecast.R
\name{fforecast}
\alias{fforecast}
\title{Functional Singular Spectrum Analysis Recurrent Forecasting and Vector Forecasting}
\usage{
fforecast(U, groups = list(1), h = 1, method = "recurrent", tol = 10^-3)
}
\arguments{
\item{U}{An object of class \code{\link{fssa}} that holds the decomposition.}

\item{groups}{A list of numeric vectors where each vector includes indices of elementary components of a group used for reconstruction and forecasting.}

\item{h}{An integer that specifies the forecast horizon.}

\item{method}{A character string specifying the type of forecasting to perform either \code{"recurrent"} or \code{"vector"}.}

\item{tol}{A double specifying the amount of tolerated error in the approximation of the matrix that corresponds with the operator formed using a Neumann series leveraged in both forecasting algorithms.}
}
\value{
A list of objects of class \code{\link{fts}} where each fts corresponds to a forecasted group.
}
\description{
This function performs functional singular spectrum analysis (FSSA) recurrent forecasting (FSSA R-forecasting) or vector forecasting (FSSA V-forecasting) of univariate or multivariate functional time series (\code{\link{fts}}) observed over a one-dimensional domain.
}
\examples{
\dontrun{
data("Callcenter")
L <- 28
U <- fssa(Callcenter, L)
groups <- list(1,1:7)
## Perform FSSA R-forecast
pr_R <- fforecast(U = U, groups = groups, h = 30, method = "recurrent")

plotly_funts(pr_R[[1]], main = "Call Center Mean Component Recurrent Forecast",
             xlab = "Time (6 minutes aggregated)",
             ylab = "Sqrt of Call Numbers",type="line",
             xticklabels = list(c("00:00","06:00","12:00","18:00","24:00")),xticklocs =
               list(c(1,60,120,180,240)))
plotly_funts(pr_R[[2]], main = "Call Center Recurrent Forecast from 7'th first components",
             xlab = "Time (6 minutes aggregated)",
             ylab = "Sqrt of Call Numbers",type="line",
             xticklabels = list(c("00:00","06:00","12:00","18:00","24:00")),xticklocs =
               list(c(1,60,120,180,240)))


## Perform FSSA V-forecast
pr_V <- fforecast(U = U, groups = groups, h = 30, method = "vector", tol = 10^-3)

plot(pr_V[[1]], mains = "Call Center Mean Component Vector Forecast",
     xlabels = "Time (6 minutes aggregated)",
     ylabels = "Sqrt of Call Numbers",type="line",
     xticklabels = list(c("00:00","06:00","12:00","18:00","24:00")),xticklocs =
       list(c(1,60,120,180,240)))
plot(pr_V[[2]], mains = "Call Center Vector Forecast from 7'th first components",
     xlabels = "Time (6 minutes aggregated)",
     ylabels = "Sqrt of Call Numbers",type="line",
     xticklabels = list(c("00:00","06:00","12:00","18:00","24:00")),xticklocs =
       list(c(1,60,120,180,240)))
}

}
