% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freconstruct.R
\name{freconstruct}
\alias{freconstruct}
\title{Reconstruction Stage of Functional Singular Spectrum Analysis}
\usage{
freconstruct(U, groups = as.list(1L:10L))
}
\arguments{
\item{U}{an object of class \code{\link{fssa}}.}

\item{groups}{a list of numeric vectors, each vector includes indices
of elementary components of a group used for reconstruction.}
}
\value{
A named list of objects of class \code{\link{funts}} that are
reconstructed according to the specified groups and a numeric vector
of eigenvalues.
}
\description{
Reconstruct univariate or multivariate functional time series
(\code{\link{funts}}) objects from functional singular spectrum analysis
(\code{\link{fssa}}) objects, including Grouping and Hankelization steps.
This function performs the reconstruction step for either univariate
functional singular spectrum analysis (ufssa) or multivariate
functional singular spectrum analysis (mfssa), depending on the input.
}
\note{
Refer to \code{\link{fssa}} for an example on how to run this
function starting from \code{\link{fssa}} objects.
}
\seealso{
\code{\link{fssa}}, \code{\link{funts}}
}
