% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfssa.R, R/freconstruct.r
\name{freconstruct}
\alias{freconstruct}
\title{Reconstruction of Functional TS}
\usage{
freconstruct(U, group = as.list(1L:10L))

freconstruct(U, group = as.list(1L:10L))
}
\arguments{
\item{U}{funtional object}

\item{U}{funtional object}
}
\value{
matrix contating Cx

matrix contating Cx
}
\description{
is a function for reconstruction stage (including Grouping and
Hankelization steps) The output is a list of functional time series corresponds to each group.
"U" in the input is a fssa object. "group" is a list.

is a function for reconstruction stage (including Grouping and
Hankelization steps) The output is a list of functional time series corresponds to each group.
'U' in the input is a fssa object. 'group' is a list.
}
