% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfssa.R, R/fssa.r
\name{fssa}
\alias{fssa}
\title{Functional SSA}
\usage{
fssa(Y, L = floor(dim(Y$coefs)[2L]/2L))

fssa(Y, L = floor(dim(Y$coefs)[2L]/2L))
}
\arguments{
\item{Y}{a functional time series.}

\item{L}{Windows length}

\item{Y}{a functional time series.}

\item{L}{Windows length}
}
\value{
list The outputs of following function is a list which
includes functional eigen vectors in the form of L-variate functional object.

list The outputs of following function is a list which
includes functional eigen vectors in the form of L-variate functional object.
}
\description{
fssa is a function for decomposition stage (including embeding
 and  SVD step) of a functional time series.

fssa is a function for decomposition stage (including embeding
 and  SVD step) of a functional time series.
}
