% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTSclass.R
\name{fts}
\alias{fts}
\title{Functional time series class}
\usage{
fts(Y, time = NULL)
}
\arguments{
\item{Y}{an object of class fd or a list of objects of class fd.}

\item{time}{the vector of times at which a time series was sampled.}
}
\description{
The function fts is used to create functional time-series objects.
}
