% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTSclass.R
\name{fts}
\alias{fts}
\title{Functional Time Series Class}
\usage{
fts(Y, time = NULL)
}
\arguments{
\item{Y}{an object of class fd or a list of objects of class fd}

\item{time}{the vector of times at which a time series was sampled}
}
\description{
This function is used to create functional time series objects from functional data (fd) objects.
}
\note{
refer to \code{\link{fssa}} for an example on how to run this function starting from fd objects
}
\seealso{
\code{\link{fssa}}
}
