% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftsplot.r
\name{ftsplot}
\alias{ftsplot}
\title{3D fts Plots with Plot3D}
\usage{
ftsplot(x, y, X, type = 2, zlab = NULL, xlab = NULL, ylab = NULL,
  space = 0.1, main = NULL)
}
\arguments{
\item{x}{a numeric vector of cordinate x.}

\item{y}{a numeric vector of cordinate y.}

\item{X}{a functional time series of class fd.}

\item{type}{what type of plot should be drawn. Possible types are
"type=1" for ribbon3D plot with curtain, "type=2" for ribbon3D plot
 without curtain, "type=3" for image2D plot.}

\item{zlab}{The lable of z axis.}

\item{xlab}{The lable of x axis.}

\item{ylab}{The lable of y axis.}

\item{space}{The amount of space (as a fraction of the average ribbon width) left between ribbons.}

\item{main}{The main title.}
}
\description{
Define some new plots for functional time series data.
}
