% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcor.r
\name{fwcor}
\alias{fwcor}
\title{Wcorrelation}
\usage{
fwcor(U, d)
}
\description{
This function evaluate the Wcorrolation plot for fssa
 @param  U in the input is a fssa object.
 @param d is the number of elementary components.
}
