% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcor.r
\name{fwcor}
\alias{fwcor}
\title{Weighted Correlation Matrix}
\usage{
fwcor(U, groups)
}
\arguments{
\item{U}{An object of class \code{\link{fssa}}.}

\item{groups}{A list or vector of indices which determines the grouping used for the reconstruction
in pairwise w-correlations matrix.}
}
\value{
A square matrix of w-correlation values for the reconstructed \code{\link{funts}} objects that were built from.
\code{\link{fssa}} components
}
\description{
This function returns the weighted correlation (w-correlation) matrix for functional time series (\code{\link{funts}}) objects
that were reconstructed from functional singular spectrum analysis (\code{\link{fssa}}) objects.
}
\examples{
\dontrun{
## Univariate FSSA Example on Callcenter data
## Univariate functional singular spectrum analysis
data("Callcenter")
L <- 28
U <- fssa(Callcenter, L)
ufwcor <- fwcor(U = U, groups = list(1, 2, 3))
wplot(W = ufwcor)
}

}
\seealso{
\code{\link{fssa}}, \code{\link{freconstruct}}, \code{\link{funts}}, \code{\link{wplot}}
}
