% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcor.r
\name{fwcor}
\alias{fwcor}
\title{Wcorrelation matrix}
\usage{
fwcor(U, d)
}
\arguments{
\item{U}{in the input is a fssa object.}

\item{d}{is the number of elementary components.}
}
\description{
This function evaluate the Wcorrolation plot for fssa
}
