% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcor.r
\name{fwcor}
\alias{fwcor}
\title{W-correlation matrix}
\usage{
fwcor(U, group)
}
\arguments{
\item{U}{in the input is an object of class \code{fssa}.}

\item{group}{A list or vector of indices determines the grouping used for the decomposition
in pairwise W-correlations matrix.}
}
\value{
A squared matrix of W-correlation values.
}
\description{
This function evaluate the Wcorrelation plot for fssa
}
\seealso{
\code{\link{fssa}}
}
