% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfssa.R, R/wcor.r
\name{fwcor}
\alias{fwcor}
\title{Wcorrelation}
\usage{
fwcor(U, d)

fwcor(U, d)
}
\description{
@param  U in the input is a fssa object.
 @param d is the number of elementary components.
 @export

@param  U in the input is a fssa object.
 @param d is the number of elementary components.
 @export
}
