% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcor.r
\name{fwcor}
\alias{fwcor}
\title{Weighted Correlation Matrix}
\usage{
fwcor(U, group)
}
\arguments{
\item{U}{an object of class \code{\link{fssa}}}

\item{group}{a list or vector of indices which determines the grouping used for the reconstruction
in pairwise w-correlations matrix}
}
\value{
a square matrix of w-correlation values
}
\description{
This function returns the weighted correlation (w-correlation) matrix for functional time series (\code{\link{fts}}) objects
that were reconstructed from functional singular spectrum analysis (\code{\link{fssa}}) objects.
}
\seealso{
\code{\link{fssa}}, \code{\link{fts}}
}
