% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launches the shinyAppDemo app}
\usage{
launchApp(type = "FSSA")
}
\arguments{
\item{type}{type of FSSA, either "FSSA" or "MFSSA"}
}
\value{
shiny application object
}
\description{
launches the shinyAppDemo app
}
\examples{
\dontrun{

launchApp()}



}
