% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_github_data.R
\docType{package}
\name{loadJambiData}
\alias{loadJambiData}
\title{Load Data from GitHub Repositories}
\usage{
loadJambiData()
}
\description{
These functions are used to load .RData files from GitHub repositories. They can be used to load the
Jambi and Montana datasets from the Rfssa package hosted on GitHub at https://github.com/haghbinh/Rfssa.
Hosting datasets on GitHub instead of including them in the Rfssa R package saves
a significant amount of space.
}
\examples{
\dontrun{
loadJambiData()
loadMontanaData()
}
}
