% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fforecast.R
\name{plot.fforecast}
\alias{plot.fforecast}
\title{Plot method for FSSA Forecast (fforecast) class}
\usage{
\method{plot}{fforecast}(
  x,
  group_index = NULL,
  ask = TRUE,
  npts = 100,
  obs = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = NULL,
  type = "l",
  lty = 1,
  ori_col = NULL,
  pred_col = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{fforecast}.}

\item{group_index}{an integer specify the group index for plot.}

\item{npts}{number of grid points for the plots.}

\item{obs}{observation number (for two-dimensional domains).}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main title for the plot.}

\item{type}{type of plot ("l" for line, "p" for points, etc.).}

\item{lty}{line type (1 for solid, 2 for dashed, etc.).}

\item{...}{additional graphical parameters passed to plotting functions.}
}
\description{
Create visualizations of FSSA Forecast (fforecast) class. This function supports
plotting `fforecast` data with one-dimensional or two-dimensional domains.
}
\seealso{
\code{\link{fforecast}}
}
