% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssaplot.R
\name{plot.fssa}
\alias{plot.fssa}
\title{Plotting fssa Objects}
\usage{
\method{plot}{fssa}(x, d = length(x$values), type = "values", ...)
}
\arguments{
\item{x}{a functional singular value decomposition object, time series objects, usually inheriting from class "fssa".}

\item{d}{an integer which is the number of elementary components in the plot.}

\item{type}{what type of plot should be drawn. Possible types are:
\itemize{
\item \code{"values"} plot the sqruare-root of singular values (default).
\item \code{"paired"} plot the pairs of eigenfunction's coefficients. (useful for the detection of periodic components).
\item \code{"wcor"} plot the W-correlation matrix for the reconstructed objects.
\item \code{"vectors"}  plot the eigenfunction's coefficients.(useful for the detection of period length).
\item \code{"meanvectors"}  plot the mean of eigenfunction's coefficients.(useful for the detection of period length).
\item \code{"meanpaired"} plot the pairs of mean of eigenfunction's coefficients. (useful for the detection of periodic components).
\item \code{"efunctions"} heatmap plot of eigenfunctions.(useful for the detection of period length).
\item \code{"efunctions2"} for plot the of eigenfunctions.(useful for the detection of meaningful patterns).
}}

\item{...}{Arguments to be passed to methods, such as graphical parameters.}
}
\description{
Plotting method for objects inheriting from class \code{\link{fssa}}.
}
\seealso{
\code{\link{fssa}}, \code{\link{ftsplot}}
}
