% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftsplot.r
\name{plot.fts}
\alias{plot.fts}
\title{Functional Time Series Plots}
\usage{
\method{plot}{fts}(Y, npts = 100, type = "line", main = "",
  ylab = NA, xlab = NA, tlab = NA, var = NA)
}
\arguments{
\item{Y}{??}

\item{npts}{??}

\item{type}{??}

\item{main}{The main title.}

\item{ylab}{The character vector of name of variables}

\item{xlab}{The label of the functions arguments.}

\item{tlab}{The time label}

\item{var}{an integer Specify the variable number.}

\item{type}{??}
}
\description{
Novel different type of plots to visualize the functional time series data objects.
}
\examples{
plot(Y,type = "heat")
plot(Y,type = "3D",var = 1)
}
