% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftsplot.r, R/plotly_funts.R
\name{plotly_funts}
\alias{plotly_funts}
\title{Functional Time Series Visualization Tools Using Plotly}
\usage{
plotly_funts(
  x,
  vars = NULL,
  types = NULL,
  subplot = TRUE,
  main = NULL,
  ylab = NULL,
  xlab = NULL,
  tlab = NULL,
  zlab = NULL,
  xticklabel = NULL,
  xticklocs = NULL,
  yticklabels = NULL,
  yticklocs = NULL,
  color_palette = "RdYlBu",
  reverse_color_palette = FALSE,
  ...
)

plotly_funts(
  x,
  vars = NULL,
  types = NULL,
  subplot = TRUE,
  main = NULL,
  ylab = NULL,
  xlab = NULL,
  tlab = NULL,
  zlab = NULL,
  xticklabel = NULL,
  xticklocs = NULL,
  yticklabels = NULL,
  yticklocs = NULL,
  color_palette = "RdYlBu",
  reverse_color_palette = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{fts}}.}

\item{vars}{Numeric specifying which variables in the FTS to plot (default: all).}

\item{types}{Tuple of strings specifying plot types for each variable.}

\item{subplot}{Logical for subplotting line plots.}

\item{main}{Titles for each plot.}

\item{ylab}{Y-axis titles.}

\item{xlab}{X-axis titles.}

\item{tlab}{Time-axis titles.}

\item{zlab}{Z-axis titles.}

\item{xticklocs}{Positions of tick labels for the domain of the functions.}

\item{yticklabels}{Tick labels for the domain of the functions.}

\item{yticklocs}{Positions of tick labels for the domain of the functions.}

\item{color_palette}{Color palette for two-dimensional FTS plots.}

\item{reverse_color_palette}{Reverse the color palette scale.}

\item{...}{Additional arguments to pass to Plotly methods.}

\item{mains}{A tuple of strings providing the the main titles of each plot.}

\item{ylabels}{A tuple of strings providing the the y-axis titles of each plot.}

\item{xlabels}{A tuple of strings providing the the x-axis titles of each plot.}

\item{tlabels}{A tuple of strings providing the the time-axis titles of each plot.}

\item{zlabels}{A tuple of strings providing the the z-axis titles of each plot.}

\item{xticklabels}{Tick labels for the domain of the functions.}
}
\description{
This is a plotting method for univariate or multivariate functional time series (\code{\link{fts}}). This method is designed to help the user visualize
\code{\link{fts}} data using a variety of techniques that use plotly.

Visualize univariate or multivariate Functional Time Series (FTS) using Plotly-based plots.
}
\details{
Supported plot types for one-dimensional domain variables:
  - "line": Line plots (default).
  - "heatmap": Heatmaps.
  - "3Dsurface": 3D surface plots.
  - "3Dline": 3D line plots.

Supported plot type for two-dimensional domain variables:
  - "heatmap"

Each variable can be plotted multiple times with different types.
}
\note{
For examples, see \code{\link{fssa}}
}
\examples{
data("Callcenter") # Univariate FTS example
plotly_funts(Callcenter, main = "Call Center Data Line Plot",
             xlab = "Time (6 minutes aggregated)",
             ylab = "Sqrt of Call Numbers", type = "line",
             xticklabels = list(c("00:00","06:00","12:00","18:00","24:00")),
             xticklocs = list(c(1,60,120,180,240)))

data("Montana") # Multivariate FTS example
plotly_funts(Montana[1:100],
             xlab = c("Time", "Longitude"),
             ylab = c("Normalized Temperature (\u00B0C)", "Latitude"),
             zlab = c("", "NDVI"),
             main = c("Temperature Curves", "NDVI Images"),
             color_palette = "RdYlGn",
             xticklabels = list(c("00:00","06:00","12:00","18:00","24:00"),
                                c("113.40\u00B0 W", "113.30\u00B0 W")),
             xticklocs = list(c(1,6,12,18,24),c(1,33)),
             yticklabels = list(NA,c("48.70\u00B0 N", "48.77\u00B0 N")),
             yticklocs = list(NA,c(1,33))
)

}
\seealso{
\code{\link{funts}}, \code{\link{Callcenter}}, \code{\link{Montana}}
}
