% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts_methods.R
\name{+.funts}
\alias{+.funts}
\title{Addition of Functional Time Series}
\usage{
\method{+}{funts}(obj1, obj2 = NULL)
}
\arguments{
\item{obj1}{an object of class \code{\link{funts}} or numeric.}

\item{obj2}{an object of class \code{\link{funts}} or numeric.}
}
\value{
an object of class \code{\link{funts}}.
}
\description{
A method for functional time series (\code{\link{funts}}) addition and funts-scalar addition.
}
\seealso{
\code{\link{funts}}
}
