% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfts_methods.R
\name{print.funts}
\alias{print.funts}
\title{Custom Print Method for Functional Time Series (funts) Objects}
\usage{
\method{print}{funts}(obj)
}
\arguments{
\item{obj}{An object of class "funts" to be printed.}
}
\description{
This function provides a custom print method for objects of the Functional Time Series (funts) class.
It displays information about the funts object, including its length (N), the number of variables,
and the structure of the object.
}
