% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts_methods.R
\name{print.funts}
\alias{print.funts}
\title{Custom Print Method for Functional Time Series (funts) Objects}
\usage{
\method{print}{funts}(x, ...)
}
\arguments{
\item{x}{an object of class "funts" to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This custom print method is designed for objects of the Functional Time Series (funts) class.
It provides a summary of the funts object, including its length (N), the number of variables,
and its structure.
}
