% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfssa.R
\docType{package}
\name{Rfssa}
\alias{Rfssa}
\alias{Rfssa-package}
\title{Rfssa: A package for Develop functional singular spectrum analysis.}
\description{
The Rfssa package provides Develop functional singular spectrum analysis
of Functional time series.
}
\section{Rfssa functions}{

The Rfssa functions: fssa, freconstruct,
 fwcor, wplot, ftsplot.
}

