% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfs_methods.R
\name{*.funts}
\alias{*.funts}
\title{Scalar Multiplication of a Functional Time Series (\code{\link{funts}}) Object}
\usage{
\method{*}{funts}(obj1, obj2)
}
\arguments{
\item{obj1}{An object of class \code{\link{funts}} or a scalar value.}

\item{obj2}{An object of class \code{\link{funts}} or a scalar value.}
}
\value{
An object of class \code{\link{funts}} representing the result of scalar multiplication.
}
\description{
Performs scalar multiplication of a Functional Time Series (funts) object by either another funts object or a scalar value.
}
\details{
This function allows you to multiply a Functional Time Series (funts) object by either another funts object or a scalar value.
The operation is element-wise when both objects are funts, and scalar multiplication when one object is a funts and the other is a scalar.
}
\seealso{
\code{\link{funts}}
}
