% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wplot.R
\name{wplot}
\alias{wplot}
\title{Weighted-Correlations Plot}
\usage{
wplot(W, cuts = 20, main = NA)
}
\arguments{
\item{W}{A w-correlation matrix.}

\item{cuts}{An integer that is the number of levels the range of w-correlation values will be divided into.}

\item{main}{A character that is the main title of the plot.}
}
\description{
This function displays a plot of the weighted-correlation (w-correlation) matrix of functional time series (\code{\link{fts}}) objects that were reconstructed
from functional singular spectrum analysis (\code{\link{fssa}}) objects.
}
\note{
Refer to \code{\link{fwcor}} for an example on how to run this function starting from a w-correlation matrix.
}
\seealso{
\code{\link{fssa}}, \code{\link{freconstruct}}, \code{\link{fts}}, \code{\link{fwcor}}
}
