% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wplot.R
\name{wplot}
\alias{wplot}
\title{Weighted-Correlations Plot}
\usage{
wplot(W, cuts = 20)
}
\arguments{
\item{W}{a w-correlation matrix}

\item{cuts}{an integer that is the number of levels the range of w-correlation values will be divided into}
}
\description{
This function displays the weighted-correlation (w-correlation) matrix of reconstructed functional time series that arise
from single component reconstructions.
}
\seealso{
\code{\link{fssa}} \code{\link{freconstruct}} \code{\link{plot.fssa}}
}
