% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wplot.R
\name{wplot}
\alias{wplot}
\title{Weighted-Correlations Plot}
\usage{
wplot(W, cuts = 20, main = NA)
}
\arguments{
\item{W}{a w-correlation matrix.}

\item{cuts}{an integer specifying the number of levels for dividing the range of w-correlation values.}

\item{main}{a character string for the main title of the plot.}
}
\description{
This function generates a plot displaying the weighted-correlation (w-correlation) matrix of functional time series (\code{\link{fts}}) objects reconstructed from functional singular spectrum analysis (\code{\link{fssa}}) objects.
}
\note{
For an example of how to use this function with a w-correlation matrix, refer to \code{\link{fwcor}}.
}
\seealso{
\code{\link{fssa}}, \code{\link{freconstruct}}, \code{\link{funts}}, \code{\link{fwcor}}
}
