% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wplot.R
\name{wplot}
\alias{wplot}
\title{Plot of W-correlations}
\usage{
wplot(W)
}
\arguments{
\item{W}{a W-correlation matrix.}
}
\description{
Display W-correlation matrix of a single components.
}
\seealso{
\code{\link{fssa}}
}
