% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfssa.R, R/wcor.r
\name{wplot}
\alias{wplot}
\title{Wplot}
\usage{
wplot(W, main = "")

wplot(W, main = "")
}
\arguments{
\item{input}{in the input is a ...}

\item{input}{in the input is a ...}
}
\description{
Wplot

Wplot
}
