% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wplot.R
\name{wplot}
\alias{wplot}
\title{Plot of W-correlations}
\usage{
wplot(W, cuts = 20)
}
\arguments{
\item{W}{a W-correlation matrix.}

\item{cuts}{integer, the number of levels the range of W-cor values will be divided into.}
}
\description{
Display W-correlation matrix of a single components.
}
\seealso{
\code{\link{fssa}}
}
